Update deprecated code and go to Python 3.6.7
This commit also fixes deprecated code found in app/urls.py. After this commit, the site *should* be compatible with Django 2.1.3.
In api_getter, it also fixes an issue where raise(e) replaces the stack of the exception. This code was changed to raise to avoid the issue.
There was also a potential security issue in cam2webui/settings.py. In the site code, the WhiteNoise middleware comes before SecurityMiddleware. In their docs, http://whitenoise.evans.io/en/stable/#quickstart-for-django-apps, they say to put the SecurityMiddleware above the WhiteNoise middleware.